avoid errors on font config

+++ hooks/xorg/post-install/50_fonts.sh
@@ -1,6 +1,7 @@
 pushd $R/etc/fonts/conf.d/
-$SUDO rm 10-*hint*
-$SUDO ln -s ../conf.avail/10-autohint.conf
-$SUDO rm 11-*lcdfilter*
-$SUDO ln -s ../conf.avail/11-lcdfilter-default.conf
+    if [ ! -e "11-lcdfilter-default.conf" ] ; then
+        $SUDO rm 10-*hint*
+        $SUDO ln -s ../conf.avail/10-autohint.conf
+        $SUDO ln -s ../conf.avail/11-lcdfilter-default.conf
+    fi
 popd
